+++ app/src/main/java/com/terra/planet/MainActivity.kt
@@ -509,6 +509,6 @@ class MainActivity : Activity() {
     }
 
     companion object {
-        const val VERSION = "0.6.0"
+        const val VERSION = "0.6.1"
     }
 }